refactor: move file operations to separate controller
