+++ src/controllers/authcontroller.js
@@ -1,5 +1,4 @@
 import * as userDb from '../db/queries/user.js';
-import * as fileDb from '../db/queries/file.js';
 import handleValidationErrors from '../utils/error.js';
 import { devLog } from '../utils/devlog.js';
 
@@ -59,22 +58,9 @@ async function handleRegister(req, res) {
     }
 }
 
-async function handleDeleteFile(req, res) {
-    const fileId = req.body.fileId;
-    try {
-        await fileDb.deleteFile(fileId);
-        devLog("File deleted:", fileId);
-        res.redirect('/dashboard');
-    } catch (error) {
-        console.error("Error deleting file:", error);
-        res.status(500).send('Error deleting file');
-    }
-}
-
 export {
     ensureAuth,
     handleLogin,
     handleLogout,
-    handleRegister,
-    handleDeleteFile,
-}
+    handleRegister
+};
