+++ src/routes/authroutes.js
@@ -5,7 +5,9 @@ import * as authController from "../controllers/authcontroller.js"
 import * as viewsController from "../controllers/viewscontroller.js"
 import * as folderController from "../controllers/foldercontroller.js"
 import * as passwordController from "../controllers/passwordcontroller.js"
+import * as fileController from "../controllers/filecontroller.js"
 import validForm from "../utils/validation.js"
+
 const router = Router()
 
 // Root & General
@@ -27,7 +29,7 @@ router.post("/register", validForm, authController.handleRegister);
 router.get("/dashboard", authController.ensureAuth, viewsController.renderDashboard);
 
 router.get("/dashboard/file/:id", authController.ensureAuth, viewsController.renderFile);
-router.post("/dashboard/file/:id", authController.ensureAuth, authController.handleDeleteFile);
+router.post("/dashboard/file/:id", authController.ensureAuth, fileController.handleDeleteFile);
 
 router.post("/dashboard/create-folder", authController.ensureAuth, folderController.handleCreateFolder);
 router.get("/dashboard/folder/:id", authController.ensureAuth, viewsController.renderFolderDashboard);
